Implement error handling for deleting a venue
Deleting a venue has been improved so that an error message is now
outputted if the venue has more than one event

+++ src/main/java/uk/ac/man/cs/eventlite/controllers/VenuesController.java
@@ -141,15 +141,25 @@ public class VenuesController {
 //    }
 //
 	@RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
-	public String deleteVenue(@PathVariable("id") long id) {
+	public String deleteVenue(@PathVariable("id") long id, RedirectAttributes redirectAttrs) {
 	
-			try {
-				venueService.delete(id);
-				return "redirect:/venues";
+		try {
+			if (venueService.findById(id).getEvents().isEmpty())
+			{
+				venueService.delete(id); 
+				return "redirect:/venues/view";
 			}
-			catch (Exception e) {
-				return "redirect:/venues";
+			else
+			{
+				System.out.println("TEST");
+				redirectAttrs.addFlashAttribute("venue_delete_failed", "A venue cannot be deleted if it has one or more (past or future) events.");
+				return "redirect:/venues/view/{id}";
 			}
+		}
+		catch (Exception e) {
+			redirectAttrs.addFlashAttribute("venue_delete_error", "An error has occurred!");
+			return "redirect:/venues";
+		}
 		
 		
 	}
